# optimize function which works 4x more fast and can work up to long 