@Jay-prakashpandey 2021-10-03
/**
* baseline method (with String.length): 2145ms
* log10 method: 711ms = 3.02 times as fast as baseline
* repeated divide: 2797ms = 0.77 times as fast as baseline
* divide-and-conquer: 74ms = 28.99 times as fast as baseline
* Mine is about 4x as fast as the divide-and-conquer solution
*/
/**
 * it computes all the powers in a loop and multiplies the ones with the corresponding set bit in n to calculate exponent in 
 *  \f$O(\log(n))\f$ using binary exponent.
 * This approach will be faster in practice sience we dont have the overhead of the recursive calls.
 */